+++ api/controllers/recipients/add-computers.js
@@ -15,15 +15,25 @@ module.exports = {
 
       // Add the recipient
       var response = await sails.helpers.recipients.add(sails.sockets.getId(this.req), this.req.payload.host, 'computers', this.req.payload.host)
+      var host = await sails.models.hosts.findOne({ host: this.req.payload.host })
 
-      // Subscribe to websockets dealing with remote audio calls for DJ Controls
+      if (!host) { return exits.error(new Error('No host found in the database')) }
+
+      // Subscribe to websockets
       if (this.req.isSocket) {
-        sails.sockets.join(this.req, 'bad-call')
-        sails.sockets.join(this.req, 'very-bad-call')
-        sails.sockets.join(this.req, 'silent-call')
-        sails.sockets.join(this.req, 'finalize-call')
-        sails.sockets.join(this.req, 'call-quality')
-        sails.log.verbose('Request was a socket. Joining bad-call and very-bad-call.')
+        if (host.makeCalls) {
+          sails.sockets.join(this.req, 'bad-call')
+          sails.sockets.join(this.req, 'very-bad-call')
+          sails.sockets.join(this.req, 'silent-call')
+          sails.sockets.join(this.req, 'finalize-call')
+        }
+        if (host.makeCalls || host.answerCalls) {
+          sails.sockets.join(this.req, 'call-quality')
+        }
+        if (host.delaySystem) {
+          sails.sockets.join(this.req, 'delay-system-dump')
+        }
+        sails.sockets.join(this.req, 'delay-system-status')
       }
 
       // Return the host label object